Module-ized prettyprinter + some small tweaks
